Further improvements

+++ backend/src/main/kotlin/com/treska/yummier/service/RecipeService.kt
@@ -1,9 +1,9 @@
 package com.treska.yummier.service
 
+import com.treska.yummier.dto.Category
+import com.treska.yummier.dto.Difficulty
 import com.treska.yummier.dto.RecipeFilter
-import com.treska.yummier.extension.normalize
-import com.treska.yummier.model.Category
-import com.treska.yummier.model.Difficulty
+import com.treska.yummier.exception.RecipeNotFoundException
 import com.treska.yummier.model.Recipe
 import com.treska.yummier.repository.RecipeRepository
 import com.treska.yummier.repository.SpecificationBuilder
@@ -12,9 +12,16 @@ import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
 
 @Service
-class RecipeService(private val recipeRepository: RecipeRepository, private val ingredientService: IngredientService) {
+class RecipeService(private val recipeRepository: RecipeRepository) {
     fun get(filter: RecipeFilter, pageable: Pageable): Page<Recipe> {
         return recipeRepository.findAll(SpecificationBuilder.withFilters(filter), pageable)
+    }
+
+    fun getById(id: Long): Recipe {
+        if (!recipeRepository.existsById(id)) {
+            throw RecipeNotFoundException("Recipe with id=[$id] does not exist.")
+        }
+        return recipeRepository.findById(id).get()
     }
 
     fun create(
@@ -24,10 +31,8 @@ class RecipeService(private val recipeRepository: RecipeRepository, private val 
         difficulty: Difficulty,
         category: Category,
         ingredients: List<String>,
-        instruction: List<String>
+        instructions: List<String>
     ): Recipe {
-        val ingredients = ingredients.map { name -> ingredientService.createIfNotExists(name.normalize()) }
-
         val recipe = Recipe(
             title = title,
             description = description,
@@ -35,7 +40,7 @@ class RecipeService(private val recipeRepository: RecipeRepository, private val 
             difficulty = difficulty,
             category = category,
             ingredients = ingredients,
-            instruction = instruction
+            instructions = instructions
         )
 
         recipeRepository.save(recipe)
@@ -44,6 +49,9 @@ class RecipeService(private val recipeRepository: RecipeRepository, private val 
     }
 
     fun delete(id: Long) {
+        if(!recipeRepository.existsById(id)) {
+            throw RecipeNotFoundException("Recipe with id=[$id] does not exist.")
+        }
         recipeRepository.deleteById(id)
     }
 }
